requirements.txt
pure-python-adb==0.3.0.dev0
uiautomator2==3.2.3
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
lxml>=5.3.0

